backend/ml-service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
torch==2.8.0
transformers==4.56.2
pillow==11.3.0
python-multipart==0.0.20
